backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
cffi==1.17.1
click==8.1.7
dnspython==2.6.1
fastapi==0.114.0
h11==0.14.0
httptools==0.6.1
idna==3.8
motor==3.5.1
pycparser==2.22
pydantic==2.9.0
pydantic-settings==2.4.0
pydantic_core==2.23.2
pymongo==4.8.0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.0.8
sniffio==1.3.1
starlette==0.38.4
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.0.1
